Some things now moved to an updater-model class. The calls there go as before via the model class. It is not tested and probably not set up as wanted later.
